Day2/requirement.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Limiter
PyJWT
pydantic
psycopg2-binary
python-dotenv